pyproject.toml
[tool.poetry]
name = "colour-profiles"
version = "0.1.0"
description = ""
authors = ["shadowlerone <andrae.lewis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
